==> scripts/06_author_histories.R <==
## This script retrieves EIDs for all of every author's prior papers
library(tidyverse)
library(xml2)

library(furrr)
library(RCurl)
source('api_key.R')

library(assertthat)
library(tictoc)

data_dir = '../data/'
author_folder = str_c(data_dir, 'author_histories/')

plan(multisession, workers = 6)


## Load and filter data ----
dropouts_03 = read_rds(str_c(data_dir, '03_dropout.Rds'))
dropouts_04 = read_rds(str_c(data_dir, '04_dropouts.Rds'))
dropouts_05 = read_rds(str_c(data_dir, '05_dropouts.Rds'))
oru_df = read_rds(str_c(data_dir, '03_matched.Rds')) %>% 
    filter(!duplicated(.)) %>% 
    anti_join(dropouts_03, by = 'auid') %>% 
    anti_join(dropouts_04, by = 'auid') %>% 
    anti_join(dropouts_05, by = 'auid')
departments = read_rds(str_c(data_dir, '03_codepartmentals.Rds'))

author_meta = read_rds(str_c(data_dir, '05_author_meta.Rds'))

anti_join(oru_df, author_meta, by = 'auid') %>% 
    nrow() %>% 
    are_equal(0L) %>% 
    assert_that(msg = 'oru_df authors missing from author_meta')


## Functions for scraping from API ----
scrape_ = function (this_auid, page_idx, max_results = 200, print_url = FALSE) {
    ## Basically just an abstraction of the RCurl call
    base_url = 'https://api.elsevier.com/content/search/scopus?'
    query_url = str_c(base_url, 
                      'query=au-id(', this_auid, ')', 'AND PUBYEAR BEF 2020',
                      '&count=', max_results,
                      '&start=', max_results*(page_idx-1),
                      '&httpAccept=application/xml',
                      '&apiKey=', api_key)
    query_url = URLencode(query_url)
    if (print_url) {
        message(query_url)
    }
    raw = getURL(query_url)
    raw
}

scrape = function (this_auid, n_docs, target_folder, 
                   force_scrape = FALSE, 
                   max_results = 200,
                   ...) {
    ## In this data, max(n_docs) = 370, so max(total_pages) = 2
    total_pages = n_docs %/% max_results + 1
    target_files = str_c(this_auid, '_', 1:total_pages, '.xml')
    for (page_idx in 1:total_pages) {
        ## Either scrape from the API + save the result OR pass
        target_file = str_c(target_folder, target_files[page_idx])
        if (!file.exists(target_file) | force_scrape) {
            raw = scrape_(this_auid, page_idx, max_results, ...)
            write_file(raw, target_file)
        }
    }
    return(target_files)
}

# ## Case w/ 2 pages
# scrape('35203386900', 202, author_folder)
# ## Case w/ incorrect count, actually only has <200 
# ## This gives a valid xml with no entries
# scrape('35498339000', 205, author_folder)

## Do the scraping ----
## ~460 sec when force_scrape = TRUE
tic()
author_history_files = author_meta %>% 
    # head(10) %>%
    # rowwise() %>% 
    future_map2(.x = .$auid, .y = .$n_docs, 
                .f = ~scrape(.x, .y, author_folder, 
                             max_results = 25,
                             force_scrape = FALSE), 
                .progress = TRUE)
toc()

## Need this for the set_names() call in the parsing workflow  
assert_that(length(author_history_files) == nrow(author_meta))

## Functions for parsing ----
parse_ = function(raw) {
    xml = read_xml(raw)
    xml = xml_ns_strip(xml)
    
    entries = xml_find_all(xml, '//entry')
    
    scopus_id = entries %>% 
        xml_find_first('dc:identifier') %>% 
        xml_text() %>% 
        str_extract('[0-9]+')
    eid = entries %>% 
        xml_find_first('eid') %>% 
        xml_text()
    doi = entries %>% 
        xml_find_first('prism:doi') %>% 
        xml_text()
    
    title = entries %>% 
        xml_find_first('dc:title') %>% 
        xml_text()
    journal = entries %>% 
        xml_find_first('prism:publicationName') %>% 
        xml_text()
    
    pub_date = entries %>% 
        xml_find_first('prism:coverDate') %>% 
        xml_text()
    
    oa = entries %>% 
        xml_find_first('openaccess') %>% 
        xml_text() %>% 
        {. == 1}
    
    cited_by = entries %>% 
        xml_find_first('citedby-count') %>% 
        xml_text() %>% 
        as.integer()
    
    tibble(scopus_id, eid, doi, 
           title, journal, 
           pub_date, 
           oa, 
           cited_by)
}
# parse_(raw)
# raw = read_file(str_c(author_folder, author_history_files[1][[1]]))
# debugonce(parse_)
# file = str_c(author_folder, author_history_files[[3]])[2]
# file %>% 
#     read_file() %>% 
#     str_remove_all('<[0-9][:word:]+>') %>% 
#     parse_()
# parse_(read_file(file))

parse = function(files, author_folder) {
    files = str_c(author_folder, files)
    parsed_df = files %>% 
        map(read_file) %>% 
        ## Remove invalid xml tags introduced by encoding errors somewhere
        map(str_remove_all, '<[0-9][:word:]+>') %>% 
        map_dfr(parse_)
    return(parsed_df)
}
# debugonce(parse)
# parse(author_history_files[[1730]], author_folder)

## Do the parsing ----
## 8.5 sec / 100 -> ~3 minutes
tic()
histories_df = author_history_files %>% 
    set_names(author_meta$auid) %>% 
    # head(100) %>%
    # {.[1730:1730]} %>%
    future_map_dfr(parse, author_folder, .id = 'auid', .progress = TRUE)
toc()

## Because of the long gap between when we pulled the metadata and downloaded the paper lists, 
## this check no longer makes sense. 
## Correct number of docs per author? 
## Except for 1 case, we have a few papers in the search that aren't in the metadata
# histories_df %>% 
#     count(auid) %>% 
#     right_join(author_meta, by = 'auid') %>% 
#     mutate(right_count = n == n_docs) %>% 
#     # filter(!right_count) %>% 
#     # count(n - n_docs)
#     # filter(abs(n - n_docs) >= 6) %>% 
#     # view()
#     ggplot(aes(log10(n), log10(n_docs))) +
#     geom_point() +
#     stat_function(fun = identity)

## How many unique papers? 
## 129k
histories_df %>% 
    pull(eid) %>% 
    unique() %>% 
    length()
## 116k using DOIs
histories_df %>% 
    filter(!is.na(doi)) %>% 
    pull(doi) %>% 
    unique() %>% 
    length()

# library(lubridate)
# ## pub_date is either ymd or NA
# # histories_df %>% 
# #     mutate(pubdate = ymd(pub_date)) %>% 
# #     filter(is.na(pubdate))
# ## Negative-cumulative count of papers by year
# ## Trimming things around 2000 -> ~90k papers
# histories_df %>% 
#     mutate(pub_date = ymd(pub_date), 
#            pub_year = year(pub_date)) %>% 
#     count(scopus_id, pub_year) %>% 
#     count(pub_year) %>% 
#     rename(n = nn) %>% 
#     mutate(cum_n = cumsum(n), 
#            cum_n_rev = max(cum_n) - cum_n) %>% 
#     ggplot(aes(pub_year, cum_n_rev)) +
#     geom_line()

## Write output ----
write_rds(histories_df, str_c(data_dir, '06_author_histories.Rds'))
